feature/OC-30 add service inspection form filled id to activity

+++ ui-library/src/main/java/br/com/stant/libraries/uilibrary/components/informationcardview/InformationCardView.java
@@ -105,7 +105,7 @@ public class InformationCardView extends CardView implements InformationCardView
     }
 
     @Override
-    public void setServiceInspectionFormFilledId(Integer serviceInspectionFormFilledId) {
+    public void setServiceInspectionFormFilledId(Long serviceInspectionFormFilledId) {
         if (serviceInspectionFormFilledId != null) {
             setWarningVisibility(View.VISIBLE);
         } else {
